Change node selectors and tolerations.
Changes the node selectors for both VICE and batch analyses to
require an 'analysis' label. The associated value does not matter.

Changes the node selectors for batch analyses to prefer nodes with
a 'batch' label. The associated value does not matter.

Changes the node selectors for VICE analyses to prefer nodes with
a VICE label. The associated value does not matter.

GPU tolerations and node selectors for VICE analyses are
unchanged.

Fixes an issue where the quota enforcement logic was blocking batch
analyses.