Use a persistent local volume for the working directory
Adds a volume claim template to the deployment for the working directory.
Make the working directory volume not dependent on the CSI driver being in use.
Add a volume mount of the volume created by the volume claim template to the analysis container.
Modifies the init container to mount the volume created by the volume claim template.

The init container will download input files to the local persistent volume.
The analysis container should mount that same persistent volume to the working directory location.